# docs/management-api-openapi.json
{
    "components": {
        "parameters": {
...
            "Limit": {
                "in": "query",
                "name": "limit",
                "schema": {
                    "default": 50,
                    "maximum": 200,
                    "minimum": 1,
                    "type": "integer"
                }
            },
            "Offset": {
                "in": "query",
                "name": "offset",
                "schema": {
                    "default": 0,
                    "minimum": 0,
                    "type": "integer"
                }
            },
...
        },
        "responses": {
            "Error400": {
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/ErrorEnvelope"
                        }
                    }
                },
                "description": "IDEMPOTENCY_KEY_REQUIRED: A task submission omitted Idempotency-Key. INVALID_CREDENTIALS: Invalid Playwright storage state or forbidden management control fields. INVALID_REQUEST: Malformed JSON, invalid field, ID, pagination or unknown setting."
            },
            "Error401": {
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/ErrorEnvelope"
                        }
                    }
                },
                "description": "UNAUTHORIZED: Missing, invalid, expired or revoked management bearer token."
            },
            "Error403": {
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/ErrorEnvelope"
                        }
                    }
                },
                "description": "CONSOLE_PASSWORD_REQUIRED: Key administration requires a fresh console-password session; model-key or old sessions do not qualify. FORBIDDEN: Authenticated key lacks a required scope."
            },
            "Error404": {
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/ErrorEnvelope"
                        }
                    }
                },
                "description": "NOT_FOUND: Resource or management route does not exist."
            },
            "Error405": {
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/ErrorEnvelope"
                        }
                    }
                },
                "description": "METHOD_NOT_ALLOWED: Known management path does not support this method; no model fallback."
            },
            "Error409": {
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/ErrorEnvelope"
                        }
                    }
                },
                "description": "ACCOUNT_BUSY: Account did not drain within 60 seconds; no forced interruption without force:true. CANCELLED: Cooperative cancellation stopped uncommitted task work. DUPLICATE_ACCOUNT: Account email already exists; import never replaces existing credentials. IDEMPOTENCY_CONFLICT: The same management key and Idempotency-Key were used with different canonical content. INTERRUPTED: Process restarted while the task was running; no automatic replay. INVALID_STATE: Requested transition is not valid in the current account/task state. VERSION_CONFLICT: Credential/state changed after verification began; stale commit rejected."
            },
            "Error413": {
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/ErrorEnvelope"
                        }
                    }
                },
                "description": "PAYLOAD_TOO_LARGE: Body exceeds 10 MiB, credentials exceed 1 MiB, or batch exceeds 100 items."
            },
            "Error429": {
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/ErrorEnvelope"
                        }
                    }
                },
                "description": "RATE_LIMITED: Management admission limit reached; retry later."
            },
            "Error500": {
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/ErrorEnvelope"
                        }
                    }
                },
                "description": "AUDIT_PERSISTENCE_FAILED: Operation committed but audit persistence failed. Read the resource/task state before retrying. INTERNAL_ERROR: Unexpected failure; message is redacted. PERSISTENCE_ERROR: Durable write failed; do not report a successful mutation. VERIFICATION_FAILED: Task item failed target-account verification; never fail over to another account. Async verification errors are reported in task/item.error; GET task still returns 200. VERIFICATION_TIMEOUT: Task item exceeded overall 10 minute verifier deadline. Async verification errors are reported in task/item.error; GET task still returns 200."
            }
        },
        "schemas": {
...
            "AuditEntry": {
                "additionalProperties": false,
                "description": "30-day retention; never store raw tokens, cookie/storage state or request credential bodies.",
                "properties": {
                    "accountId": {
                        "description": "Opaque stable identifier; never use the filesystem index as an account ID.",
                        "minLength": 1,
                        "type": "string"
                    },
                    "action": {
                        "type": "string"
                    },
                    "auditId": {
                        "description": "Opaque stable identifier; never use the filesystem index as an account ID.",
                        "minLength": 1,
                        "type": "string"
                    },
                    "createdByKeyId": {
                        "description": "Opaque stable identifier; never use the filesystem index as an account ID.",
                        "minLength": 1,
                        "type": "string"
                    },
                    "error": {
                        "$ref": "#/components/schemas/Error"
                    },
                    "outcome": {
                        "type": "string"
                    },
                    "requestId": {
                        "type": "string"
                    },
                    "taskId": {
                        "description": "Opaque stable identifier; never use the filesystem index as an account ID.",
                        "minLength": 1,
                        "type": "string"
                    },
                    "timestamp": {
                        "format": "date-time",
                        "type": "string"
                    }
                },
                "required": ["auditId", "requestId", "createdByKeyId", "action", "timestamp", "outcome"],
                "type": "object"
            },
            "AuditPage": {
                "additionalProperties": false,
                "properties": {
                    "items": {
                        "items": {
                            "$ref": "#/components/schemas/AuditEntry"
                        },
                        "type": "array"
                    },
                    "limit": {
                        "maximum": 200,
                        "minimum": 1,
                        "type": "integer"
                    },
                    "offset": {
                        "minimum": 0,
                        "type": "integer"
                    },
                    "total": {
                        "minimum": 0,
                        "type": "integer"
                    }
                },
                "required": ["items", "total", "offset", "limit"],
                "type": "object"
            },
            "AuditPageEnvelope": {
                "additionalProperties": false,
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/AuditPage"
                    },
                    "requestId": {
                        "type": "string"
                    }
                },
                "required": ["data", "requestId"],
                "type": "object"
            },
...
            "Error": {
                "additionalProperties": false,
                "properties": {
                    "code": {
                        "$ref": "#/components/schemas/ErrorCode"
                    },
                    "message": {
                        "type": "string"
                    }
                },
                "required": ["code", "message"],
                "type": "object"
            },
            "ErrorCode": {
                "description": "HTTP mappings and task-only codes are documented in x-error-catalog.",
                "enum": [
                    "INVALID_REQUEST",
                    "INVALID_CREDENTIALS",
                    "IDEMPOTENCY_KEY_REQUIRED",
                    "UNAUTHORIZED",
                    "FORBIDDEN",
                    "CONSOLE_PASSWORD_REQUIRED",
                    "NOT_FOUND",
                    "METHOD_NOT_ALLOWED",
                    "IDEMPOTENCY_CONFLICT",
                    "DUPLICATE_ACCOUNT",
                    "VERSION_CONFLICT",
                    "ACCOUNT_BUSY",
                    "INVALID_STATE",
                    "PAYLOAD_TOO_LARGE",
                    "RATE_LIMITED",
                    "PERSISTENCE_ERROR",
                    "INTERNAL_ERROR",
                    "VERIFICATION_FAILED",
                    "VERIFICATION_TIMEOUT",
                    "CANCELLED",
                    "INTERRUPTED",
                    "AUDIT_PERSISTENCE_FAILED"
                ],
                "type": "string"
            },
            "ErrorEnvelope": {
                "additionalProperties": false,
                "properties": {
                    "error": {
                        "$ref": "#/components/schemas/Error"
                    },
                    "requestId": {
                        "type": "string"
                    }
                },
                "required": ["error", "requestId"],
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "ConsoleSession": {
                "description": "Console-password session; session.authMethod must equal console_password. Model-key login and legacy sessions require console password re-login.",
                "in": "cookie",
                "name": "connect.sid",
                "type": "apiKey"
            },
            "ManagementBearer": {
                "bearerFormat": "mgmt_<32-random-bytes-encoded>",
                "description": "Management token only. No console session or model API key fallback. SHA-256 hash persisted; plaintext once.",
                "scheme": "bearer",
                "type": "http"
            }
        }
    },
    "info": {
        "description": "External management API validated by real-service local simulated acceptance. Actual ProxyServerSystem mounts, storage, key/task/account services, verifier, loopback WebSocket and real client script are exercised; browser resources and upstream fetch are synthetic. LIVE-01 is explicitly deferred and is not a pass. Single process/replica; offline CLI requires a stopped service. No deployment or real credentials/model calls in this round.",
        "title": "Aitoapi external management API",
        "version": "1.0.0"
    },
    "openapi": "3.0.3",
    "paths": {
...
        "/api/manage/v1/audit": {
            "get": {
                "description": "Paginated redacted audit records retained for 30 days.",
                "operationId": "listAudit",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/Offset"
                    },
                    {
                        "$ref": "#/components/parameters/Limit"
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuditPageEnvelope"
                                }
                            }
                        },
                        "description": "Success."
                    },
                    "400": {
                        "$ref": "#/components/responses/Error400"
                    },
                    "401": {
                        "$ref": "#/components/responses/Error401"
                    },
                    "403": {
                        "$ref": "#/components/responses/Error403"
                    },
                    "404": {
                        "$ref": "#/components/responses/Error404"
                    },
                    "405": {
                        "$ref": "#/components/responses/Error405"
                    },
                    "409": {
                        "$ref": "#/components/responses/Error409"
                    },
                    "413": {
                        "$ref": "#/components/responses/Error413"
                    },
                    "429": {
                        "$ref": "#/components/responses/Error429"
                    },
                    "500": {
                        "$ref": "#/components/responses/Error500"
                    }
                },
                "security": [
                    {
                        "ManagementBearer": []
                    }
                ],
                "summary": "listAudit",
                "tags": ["audit"],
                "x-required-scopes": ["audit:read"],
                "x-response-mode": "sync"
            }
        },
...
    },
    "security": [
        {
            "ManagementBearer": []
        }
    ],
    "servers": [
        {
            "description": "Same origin as the deployed service; examples use an isolated localhost server.",
            "url": "/"
        }
    ],
...
}